arm: tegra13: Always enfore 2-core CPU EDP limit
On T132, offlining a core within Linux is not sufficient to guarantee that
the core is continually in a powered-down state. Until we do have a sufficient
guarantee, we will use EDP tables to always enforce the 2-core freq cap even
when a single core is ONLINE.

To do this in the simplest way, values in the dynamic_constants[] array and
the leakage_constants[] array are all set to the 2-core value below in the
structure init. With this, there's no need to modify the code edp logic in
edp.c and it remains specific to this chip.

The actual 1-core values for these two arrays is saved as a comment to
resurrect when actually needed.

Bug 200093655

Change-Id: Ibbca382807da0f7146ad9af361581eceb03976e9
Signed-off-by: Nine Feng <nfeng@nvidia.com>
Reviewed-on: http://git-master/r/764278